% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_hscode.R
\name{parse_hscode}
\alias{parse_hscode}
\title{convert hscode to Chinese name or English name}
\usage{
parse_hscode(x, locale = "Chinese")
}
\arguments{
\item{x}{a character vector hscode}

\item{locale}{Chinese name or English}
}
\value{
a character vector
}
\description{
convert hscode to Chinese name or English name
}
\examples{
parse_hscode(c("01", "02", "0101290000"))
}
