% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_hscode.R
\name{search_hscode}
\alias{search_hscode}
\title{Search HS Code Chinese name}
\usage{
search_hscode(hscode = NULL, chinese = NULL)
}
\arguments{
\item{hscode}{a character, 2 - 11 digits hscode}

\item{chinese}{Chinese key word}
}
\value{
data.frame
}
\description{
Search HS Code Chinese name
}
\examples{
search_hscode("0101")
}
